Only import public methods via contract/__init__
