+++ tests/core/contracts/test_contract_util_functions.py
@@ -1,13 +1,11 @@
 import pytest
 
 from web3._utils.contracts import (
-    validate_payable,
-)
-from web3.contract import (
     async_parse_block_identifier,
     async_parse_block_identifier_int,
     parse_block_identifier,
     parse_block_identifier_int,
+    validate_payable,
 )
 from web3.exceptions import (
     BlockNumberOutofRange,
